pyproject.toml
[tool.poetry]
name = "dica-scraper2"
version = "0.1.0"
description = ""
authors = ["Joshua Westphal Da Silva"]

[tool.poetry.dependencies]
python = "^3.10"
pytest-playwright = "^0.3.0"
ipython = "^8.4.0"
beautifulsoup4 = "^4.11.1"
pandas = "^1.4.3"
typer = {version = "^0.6.1", extras = ["rich shellingham"]}

[tool.poetry.dev-dependencies]
black = "^22.6.0"
rich = "^12.5.1"
ipython = "^8.4.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
